remove channel should always redirect even if removal fails

+++ hc/front/views.py
@@ -336,11 +336,12 @@ def verify_email(request, code, token):
 def remove_channel(request, code):
     assert request.method == "POST"
 
-    channel = get_object_or_404(Channel, code=code)
-    if channel.user != request.user:
-        return HttpResponseForbidden()
-
-    channel.delete()
+    # user may refresh the page during POST and cause two deletion attempts
+    channel = Channel.objects.filter(code=code).first()
+    if channel:
+        if channel.user != request.user:
+            return HttpResponseForbidden()
+        channel.delete()
 
     return redirect("hc-channels")
